requirements.txt
fastapi~=0.104.0
uvicorn==0.23.2
asyncpg==0.28.0

SQLAlchemy~=2.0.22
pydantic~=2.4.2
pydantic_settings~=2.0.3
alembic~=1.12.1
passlib~=1.7.4
python-jose==3.3.0
pytz~=2023.3.post1
pydantic[email]
python-multipart
